Import of RMBARKER/File-Copy-Link-0.100 from CPAN.
gitpan-cpan-distribution: File-Copy-Link
gitpan-cpan-version:      0.100
gitpan-cpan-path:         RMBARKER/File-Copy-Link-0.100.tar.gz
gitpan-cpan-author:       RMBARKER
gitpan-cpan-maturity:     released

+++ t/copylink.t
@@ -1,12 +1,17 @@
+#!perl
 # Before `make install' is performed this script should be runnable with
 # `make test'. After `make install' it should work as `perl copylink.t'
+
+use strict;
+use warnings;
+our $VERSION = 1.0;
 
 #########################
 
 # change 'tests => 1' to 'tests => last_test_to_print';
 
 use Test::More tests => 6;
-BEGIN { use_ok('File::Copy::Link') };
+BEGIN { use_ok('File::Copy::Link', qw(copylink) ) };
 
 #########################
 
@@ -14,44 +19,45 @@ BEGIN { use_ok('File::Copy::Link') };
 # its man page ( perldoc Test::More ) for help writing this test script.
 
 use File::Compare;
-use File::Path;
+use File::Temp qw(tempdir);
 
-my $dir = 'test';
-if( -e $dir ) { rmtree $dir or die }
-mkpath $dir or die;
+my $dir = tempdir;
 
 my $file = File::Spec->catfile($dir,'file.txt');
 my $link = File::Spec->catfile($dir,'link.lnk');
 
-open my $fh, ">", $file or die;
-print $fh "text\n" or die;
+open my $fh, q{>}, $file or die;
+print {$fh} "text\n" or die;
 close $fh or die;
 
 SKIP: {
-    skip "'symlink' not implemented", 5 unless eval{ symlink("",""); 1 }; 
+    if( !eval{ (symlink q{}, q{}), 1 } ) { 
+	skip q{'symlink' not implemented}, 5 
+    }
+    # die if not((symlink 'file.txt', $link) and (-l $link) and not(compare($file,$link)));
+    die $! if not(symlink 'file.txt', $link);
+    die if not(-l $link); 
+    die if compare($file,$link);
 
-    die unless
-	symlink('file.txt',$link) && -l $link && !compare($file,$link);
-
-    open $fh, ">>", $file or die;
-    print $fh "more\n" or die;
+    open $fh, q{>>}, $file or die;
+    print {$fh} "more\n" or die;
     close $fh or die;
-    !compare($file,$link) or die;
+    not compare($file,$link) or die;
 
-    ok( copylink($link), "copylink");
-    ok( !(-l $link), "not a link");
-    ok( !compare($file,$link), "compare file and copy");
+    ok( copylink($link), q{copylink});
+    ok( !(-l $link), q{not a link});
+    ok( !compare($file,$link), q{compare file and copy});
 
-    open $fh, ">>", $file or die;
-    print $fh "more\n" or die;
+    open $fh, q{>>}, $file or die;
+    print {$fh} qq{more\n} or die;
     close $fh or die;
 
     compare($file,$link) or die;
     unlink $file or die;
 
-    ok( -e $link, "copy not deleted"); 
+    ok( -e $link, q{copy not deleted}); 
     unlink $link or die;
-    ok( !(-e $link), "copy deleted");
+    ok( !(-e $link), q{copy deleted});
 }
 
-END { rmtree $dir if $dir };
+# $Id: copylink.t 82 2006-07-26 08:55:37Z rmb1 $
